requirements.txt
APScheduler==3.10.4
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
